Citizen_Grievance_AI/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
nltk==3.8.1
textblob==0.17.1
geopy==2.4.1
pillow==10.1.0
transformers==4.36.2
torch==2.1.2
speechrecognition==3.10.1
